Time: 5 ms (39.96%) | Memory: 69.1 MB (96.91%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -10,6 +10,6 @@ class Solution {
             sum +=nums[i]-nums[i-k];
             maxsum=Math.max(maxsum,sum);
         }
-        return (double) maxsum/k;
+        return maxsum/k;
     }
 }
